+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/config/MapperConfig.java
@@ -17,10 +17,11 @@
 
 package org.apache.shenyu.admin.config;
 
+import org.apache.shenyu.admin.mybatis.common.SqlUpdateInterceptor;
 import org.apache.shenyu.admin.mybatis.oracle.OracleSQLPrepareInterceptor;
 import org.apache.shenyu.admin.mybatis.pg.interceptor.PostgreSQLPrepareInterceptor;
 import org.apache.shenyu.admin.mybatis.pg.interceptor.PostgreSQLQueryInterceptor;
-import org.apache.shenyu.admin.mybatis.pg.interceptor.PostgreSqlUpdateInterceptor;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -68,22 +69,12 @@ public class MapperConfig {
     /**
      * Add the plugin to the MyBatis plugin interceptor chain.
      *
-     * @return {@linkplain PostgreSqlUpdateInterceptor}
+     * @return {@linkplain SqlUpdateInterceptor}
      */
     @Bean
-    @ConditionalOnProperty(name = "shenyu.database.dialect", havingValue = "oracle")
-    public PostgreSqlUpdateInterceptor oracleSqlUpdateInterceptor() {
-        return new PostgreSqlUpdateInterceptor();
+    @ConditionalOnExpression(value = "'${shenyu.database.dialect}' == 'oracle' || '${shenyu.database.dialect}' == 'postgresql'")
+    public SqlUpdateInterceptor oracleSqlUpdateInterceptor() {
+        return new SqlUpdateInterceptor();
     }
 
-    /**
-     * Add the plugin to the MyBatis plugin interceptor chain.
-     *
-     * @return {@linkplain PostgreSqlUpdateInterceptor}
-     */
-    @Bean
-    @ConditionalOnProperty(name = "shenyu.database.dialect", havingValue = "postgresql")
-    public PostgreSqlUpdateInterceptor postgreSqlUpdateInterceptor() {
-        return new PostgreSqlUpdateInterceptor();
-    }
 }
